Fix implicit conversion error

+++ src/app/clusters/color-control-server/ColorControlCluster.cpp
@@ -366,7 +366,7 @@ bool ColorControlCluster::TickHue(HueTransition & tx, uint64_t now)
         const float t     = std::clamp(static_cast<float>(elapsed) / static_cast<float>(tx.durationMs), 0.f, 1.f);
         done              = (t >= 1.f);
         const int32_t arc = done ? tx.signedDelta                            // exact endpoint, no drift
-                                 : static_cast<int32_t>(tx.signedDelta * t); // 16-bit circular interpolation
+                                 : static_cast<int32_t>(static_cast<float>(tx.signedDelta) * t); // 16-bit circular interpolation
         eh                = static_cast<uint16_t>((static_cast<int32_t>(tx.startHue) + arc) & 0xFFFF);
     }
 
@@ -405,7 +405,9 @@ bool ColorControlCluster::TickSat(SatTransition & tx, uint64_t now)
 
     // Saturation is linear (not circular) and 8-bit. Exact target on the last tick avoids drift.
     const uint8_t sat =
-        done ? tx.targetSat : static_cast<uint8_t>(tx.startSat + static_cast<int32_t>((int32_t(tx.targetSat) - tx.startSat) * t));
+        done ? tx.targetSat
+             : static_cast<uint8_t>(tx.startSat +
+                                    static_cast<int32_t>(static_cast<float>(int32_t(tx.targetSat) - tx.startSat) * t));
 
     // NETWORK REPORTING is throttled (§3.2.7.2): silent mid-transition, report at the end.
     const auto change = done ? AttributeChangeType::kReportable : AttributeChangeType::kQuiet;
@@ -440,7 +442,8 @@ bool ColorControlCluster::TickCT(CTTransition & tx, uint64_t now)
     // Color temperature is linear, 16-bit mireds. Exact target on the last tick avoids drift.
     const uint16_t mireds = done
         ? tx.targetMireds
-        : static_cast<uint16_t>(tx.startMireds + static_cast<int32_t>((int32_t(tx.targetMireds) - tx.startMireds) * t));
+        : static_cast<uint16_t>(tx.startMireds +
+                                static_cast<int32_t>(static_cast<float>(int32_t(tx.targetMireds) - tx.startMireds) * t));
 
     // NETWORK REPORTING is throttled (§3.2.7.2): silent mid-transition, report at the end.
     const auto change = done ? AttributeChangeType::kReportable : AttributeChangeType::kQuiet;
@@ -520,9 +523,11 @@ bool ColorControlCluster::TickXY(XYTransition & tx, uint64_t now)
 
     // CIE X/Y are 16-bit and linear. Exact target on each axis's last tick avoids drift.
     const uint16_t x =
-        done ? tx.targetX : static_cast<uint16_t>(tx.startX + static_cast<int32_t>((int32_t(tx.targetX) - tx.startX) * tX));
+        done ? tx.targetX
+             : static_cast<uint16_t>(tx.startX + static_cast<int32_t>(static_cast<float>(int32_t(tx.targetX) - tx.startX) * tX));
     const uint16_t y =
-        done ? tx.targetY : static_cast<uint16_t>(tx.startY + static_cast<int32_t>((int32_t(tx.targetY) - tx.startY) * tY));
+        done ? tx.targetY
+             : static_cast<uint16_t>(tx.startY + static_cast<int32_t>(static_cast<float>(int32_t(tx.targetY) - tx.startY) * tY));
 
     // NETWORK REPORTING is throttled (3.2.7.2): silent mid-transition, report only when BOTH done.
     const auto change = done ? AttributeChangeType::kReportable : AttributeChangeType::kQuiet;
